Fix platform tests for new matviews
