+++ test/platform-checks/mzcompose.py
@@ -29,10 +29,10 @@ from materialize.checks.join_types import *  # noqa: F401 F403
 from materialize.checks.jsonb_type import *  # noqa: F401 F403
 from materialize.checks.large_tables import *  # noqa: F401 F403
 from materialize.checks.like import *  # noqa: F401 F403
+from materialize.checks.materialized_views import *  # noqa: F401 F403
 from materialize.checks.nested_types import *  # noqa: F401 F403
 from materialize.checks.null_value import *  # noqa: F401 F403
 from materialize.checks.numeric_types import *  # noqa: F401 F403
-from materialize.checks.recorded_views import *  # noqa: F401 F403
 from materialize.checks.regex import *  # noqa: F401 F403
 from materialize.checks.rename_index import *  # noqa: F401 F403
 from materialize.checks.rename_source import *  # noqa: F401 F403
